Add silverkeyinc.com
```
I am working with an awesome startup in San Francisco that just got "Music app of the year" in the app store. There were founded in 2010 and have just closed an A round of funding for $4.75M. They are now looking for some good people that are wanting to get in on the ground floor and help grow a company into something amazing.

From looking at your profile you look the kind of person they are looking for. Are you looking for new opportunities?
```